authorization: Move AccessRuleList to its own module
